requirements.txt
panel
holoviews
hvplot
streamlit_option_menu
streamlit-aggrid
